Fix warning: result of comparison of constant -1 with expression of type 'uint16_t' is always true Fix warning: illegal character encoding in string literal

+++ src/level/level_light.cpp
@@ -157,7 +157,7 @@ IntSector *FLevel::GetSectorFromSubSector(const MapSubsectorEx *sub)
 	for (int i = 0; i < (int)sub->numlines; i++)
 	{
 		MapSegGLEx *seg = &GLSegs[sub->firstline + i];
-		if (seg->side != NO_SIDE_INDEX)
+		if ((int16_t)seg->side != NO_SIDE_INDEX)
 		{
 			IntLineDef *line = &Lines[seg->linedef];
 			return GetFrontSector(&Sides[line->sidenum[seg->side]]);
